refactor: consolidate remaining crypto primitives in barretenberg/ts
Move all remaining standard cryptographic operations from foundation/crypto
to barretenberg/ts, continuing the consolidation started in previous commits.

Moved modules:
- AES-128-CBC encryption/decryption
- BN254 G1/G2 operations with point compression
- Secp256k1 elliptic curve operations
- Verification key utilities (vkAsFieldsMegaHonk)

Foundation modules now serve as thin adapters that:
- Re-export or delegate to @aztec/bb.js/crypto/* subpaths
- Maintain backward compatibility with Buffer return types
- Convert between foundation Fr and barretenberg Bn254Fr types where needed

This completes the crypto consolidation effort, with all standard
cryptographic primitives now centralized in barretenberg/ts.
