+++ barretenberg/ts/src/crypto/aes128/index.ts
@@ -1,0 +1,72 @@
+/**
+ * AES-128-CBC encryption/decryption using barretenberg bbapi.
+ */
+
+import { BarretenbergSync } from '../../barretenberg/index.js';
+
+/**
+ * AES-128-CBC encryption/decryption.
+ */
+export class Aes128 {
+  /**
+   * Encrypt a buffer using AES-128-CBC.
+   * @param data - Data to encrypt.
+   * @param iv - AES initialization vector.
+   * @param key - Key to encrypt with.
+   * @returns Encrypted data.
+   */
+  public async encryptBufferCBC(data: Uint8Array, iv: Uint8Array, key: Uint8Array) {
+    const rawLength = data.length;
+    const numPaddingBytes = 16 - (rawLength % 16);
+    const paddingBuffer = new Uint8Array(numPaddingBytes);
+    // input num bytes needs to be a multiple of 16 and at least 1 byte
+    // node uses PKCS#7-Padding scheme, where padding byte value = the number of padding bytes
+    paddingBuffer.fill(numPaddingBytes);
+    const input = new Uint8Array(data.length + paddingBuffer.length);
+    input.set(data);
+    input.set(paddingBuffer, data.length);
+
+    await BarretenbergSync.initSingleton();
+    const api = BarretenbergSync.getSingleton();
+    const response = api.aesEncrypt({
+      plaintext: input,
+      iv,
+      key,
+      length: input.length,
+    });
+    return new Uint8Array(response.ciphertext);
+  }
+
+  /**
+   * Decrypt a buffer using AES-128-CBC.
+   * We keep the padding in the returned buffer.
+   * @param data - Data to decrypt.
+   * @param iv - AES initialization vector.
+   * @param key - Key to decrypt with.
+   * @returns Decrypted data.
+   */
+  public async decryptBufferCBCKeepPadding(data: Uint8Array, iv: Uint8Array, key: Uint8Array): Promise<Uint8Array> {
+    await BarretenbergSync.initSingleton();
+    const api = BarretenbergSync.getSingleton();
+    const response = api.aesDecrypt({
+      ciphertext: data,
+      iv,
+      key,
+      length: data.length,
+    });
+    return new Uint8Array(response.plaintext);
+  }
+
+  /**
+   * Decrypt a buffer using AES-128-CBC.
+   * @param data - Data to decrypt.
+   * @param iv - AES initialization vector.
+   * @param key - Key to decrypt with.
+   * @returns Decrypted data.
+   */
+  public async decryptBufferCBC(data: Uint8Array, iv: Uint8Array, key: Uint8Array) {
+    const paddedBuffer = await this.decryptBufferCBCKeepPadding(data, iv, key);
+    const paddingToRemove = paddedBuffer[paddedBuffer.length - 1];
+    return paddedBuffer.subarray(0, paddedBuffer.length - paddingToRemove);
+  }
+}
