+++ yarn-project/foundation/src/crypto/poseidon/index.ts
@@ -1,4 +1,16 @@
-import { BarretenbergSync } from '@aztec/bb.js';
+/**
+ * Poseidon2 hash operations - delegates to barretenberg/ts.
+ * This wrapper maintains Fr return types for backward compatibility.
+ */
+import {
+  type Fieldable as FieldableBb,
+  poseidon2HashAccumulate as poseidon2HashAccumulateImpl,
+  poseidon2HashBytes as poseidon2HashBytesImpl,
+  poseidon2Hash as poseidon2HashImpl,
+  poseidon2HashWithSeparator as poseidon2HashWithSeparatorImpl,
+  poseidon2Permutation as poseidon2PermutationImpl,
+} from '@aztec/bb.js/crypto/poseidon';
+import { Bn254Fr } from '@aztec/bb.js/types/fields';
 
 import { Fr } from '../../fields/fields.js';
 import { type Fieldable, serializeToFields } from '../../serialize/serialize.js';
@@ -10,12 +22,10 @@ import { type Fieldable, serializeToFields } from '../../serialize/serialize.js'
  */
 export async function poseidon2Hash(input: Fieldable[]): Promise<Fr> {
   const inputFields = serializeToFields(input);
-  await BarretenbergSync.initSingleton();
-  const api = BarretenbergSync.getSingleton();
-  const response = api.poseidon2Hash({
-    inputs: inputFields.map(i => i.toBuffer()),
-  });
-  return Fr.fromBuffer(Buffer.from(response.hash));
+  // Convert Fr to Bn254Fr for barretenberg
+  const bbFields = inputFields.map(f => Bn254Fr.fromBuffer(f.toBuffer())) as FieldableBb[];
+  const result = await poseidon2HashImpl(bbFields);
+  return Fr.fromBuffer(Buffer.from(result.toBuffer()));
 }
 
 /**
@@ -26,23 +36,21 @@ export async function poseidon2Hash(input: Fieldable[]): Promise<Fr> {
  */
 export async function poseidon2HashWithSeparator(input: Fieldable[], separator: number): Promise<Fr> {
   const inputFields = serializeToFields(input);
-  inputFields.unshift(new Fr(separator));
-  await BarretenbergSync.initSingleton();
-  const api = BarretenbergSync.getSingleton();
-  const response = api.poseidon2Hash({
-    inputs: inputFields.map(i => i.toBuffer()),
-  });
-  return Fr.fromBuffer(Buffer.from(response.hash));
+  const bbFields = inputFields.map(f => Bn254Fr.fromBuffer(f.toBuffer())) as FieldableBb[];
+  const result = await poseidon2HashWithSeparatorImpl(bbFields, separator);
+  return Fr.fromBuffer(Buffer.from(result.toBuffer()));
 }
 
+/**
+ * Create a poseidon2 hash accumulate from an array of input fields.
+ * @param input - The input fields to hash.
+ * @returns The poseidon2 hash.
+ */
 export async function poseidon2HashAccumulate(input: Fieldable[]): Promise<Fr> {
   const inputFields = serializeToFields(input);
-  await BarretenbergSync.initSingleton();
-  const api = BarretenbergSync.getSingleton();
-  const response = api.poseidon2HashAccumulate({
-    inputs: inputFields.map(i => i.toBuffer()),
-  });
-  return Fr.fromBuffer(Buffer.from(response.hash));
+  const bbFields = inputFields.map(f => Bn254Fr.fromBuffer(f.toBuffer())) as FieldableBb[];
+  const result = await poseidon2HashAccumulateImpl(bbFields);
+  return Fr.fromBuffer(Buffer.from(result.toBuffer()));
 }
 
 /**
@@ -52,34 +60,17 @@ export async function poseidon2HashAccumulate(input: Fieldable[]): Promise<Fr> {
  */
 export async function poseidon2Permutation(input: Fieldable[]): Promise<Fr[]> {
   const inputFields = serializeToFields(input);
-  // We'd like this assertion but it's not possible to use it in the browser.
-  // assert(input.length === 4, 'Input state must be of size 4');
-  await BarretenbergSync.initSingleton();
-  const api = BarretenbergSync.getSingleton();
-  const response = api.poseidon2Permutation({
-    inputs: inputFields.map(i => i.toBuffer()),
-  });
-  // We'd like this assertion but it's not possible to use it in the browser.
-  // assert(response.outputs.length === 4, 'Output state must be of size 4');
-  return response.outputs.map(o => Fr.fromBuffer(Buffer.from(o)));
+  const bbFields = inputFields.map(f => Bn254Fr.fromBuffer(f.toBuffer())) as FieldableBb[];
+  const results = await poseidon2PermutationImpl(bbFields);
+  return results.map(o => Fr.fromBuffer(Buffer.from(o.toBuffer())));
 }
 
+/**
+ * Create a poseidon2 hash from a byte buffer.
+ * @param input - The input bytes to hash.
+ * @returns The poseidon2 hash.
+ */
 export async function poseidon2HashBytes(input: Buffer): Promise<Fr> {
-  const inputFields = [];
-  for (let i = 0; i < input.length; i += 31) {
-    const fieldBytes = Buffer.alloc(32, 0);
-    input.slice(i, i + 31).copy(fieldBytes);
-
-    // Noir builds the bytes as little-endian, so we need to reverse them.
-    fieldBytes.reverse();
-    inputFields.push(Fr.fromBuffer(fieldBytes));
-  }
-
-  await BarretenbergSync.initSingleton();
-  const api = BarretenbergSync.getSingleton();
-  const response = api.poseidon2Hash({
-    inputs: inputFields.map(i => i.toBuffer()),
-  });
-
-  return Fr.fromBuffer(Buffer.from(response.hash));
+  const result = await poseidon2HashBytesImpl(input);
+  return Fr.fromBuffer(Buffer.from(result.toBuffer()));
 }
